@@ -18,75 +18,87 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.util.StringUtil;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
-public final class Main implements CommandExecutor, TabCompleter {
+public final class MainCommand implements CommandExecutor, TabCompleter {
 
     private final RoulettePlugin plugin;
 
-    public Main(RoulettePlugin plugin) {
+    private static final List<String> COMMAND_ARGS = List.of("create", "delete", "reload", "map");
+    private static final List<String> TABLE_NAME_ARG = List.of("<name>");
+    private static final List<String> TYPES = List.of("american", "european");
+    private static final List<String> HELP = Stream.of(
+            "&8&m--------------------------------------------------",
+            "&6&lRoulette &f&oCommands &c(optional) <required>",
+            "&e/roulette create <name> <type> &f- &7Create a new roulette.",
+            "&e/roulette delete <name> &f- &7Delete a game.",
+            "&e/roulette reload &f- &7Reload configuration files.",
+            "&e/roulette map &f- &7Gives a win voucher.",
+            "&8&m--------------------------------------------------").map(PluginUtils::translate).toList();
+
+    public MainCommand(RoulettePlugin plugin) {
         this.plugin = plugin;
     }
 
     @Override
     public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String label, @NotNull String[] args) {
+        // If player doesn't have permission to use roulette commands, send @no-permission message.
+        if (!hasPermission(sender, "roulette.help")) return true;
+
+        MessageManager messages = plugin.getMessageManager();
+
         // This command can't be executed from the console.
-        if (!(sender instanceof Player)) {
-            plugin.getMessageManager().send(sender, MessageManager.Message.FROM_CONSOLE);
+        if (!(sender instanceof Player player)) {
+            messages.send(sender, MessageManager.Message.FROM_CONSOLE);
             return true;
         }
 
-        Player player = (Player) sender;
-
         // No arguments provided.
-        if (args.length == 0 || args.length > 3) {
-            // If player doesn't have permission to use roulette commands, send @no-permission message.
-            if (!hasPermission(player, "roulette.help")) return true;
-
+        boolean noArgs = args.length == 0;
+        if (noArgs || args.length > 3 || !COMMAND_ARGS.contains(args[0].toLowerCase())) {
             // Otherwise, send help message.
-            plugin.getMessageManager().send(player, args.length == 0 ? MessageManager.Message.HELP : MessageManager.Message.SINTAX);
+            if (noArgs) HELP.forEach(player::sendMessage);
+            else messages.send(player, MessageManager.Message.SINTAX);
             return true;
         }
 
         if (args.length == 1) {
             switch (args[0].toLowerCase()) {
-                case "reload":
+                case "reload" -> {
                     // If player doesn't have permission to reload, send @no-permission message.
                     if (!hasPermission(player, "roulette.reload")) return true;
 
                     // Log reloading message.
                     plugin.getLogger().info("Reloading " + plugin.getDescription().getFullName());
 
-                    // Reload main config.
-                    plugin.reloadConfig();
-                    plugin.reloadAbbreviations();
+                    messages.send(sender, MessageManager.Message.RELOADING);
+                    CompletableFuture.runAsync(plugin::updateConfigs).thenRun(() -> plugin.getServer().getScheduler().runTask(plugin, () -> {
+                        plugin.reloadAbbreviations();
 
-                    // Reload winners config.
-                    plugin.getWinnerManager().reloadConfig();
+                        // Reload winners config.
+                        plugin.getWinnerManager().reloadConfig();
 
-                    // Reload chips config.
-                    plugin.getChipManager().reloadConfig();
+                        // Reload chips config.
+                        plugin.getChipManager().reloadConfig();
 
-                    // Reload messages config.
-                    plugin.getMessageManager().reloadConfig();
+                        // Reload games.
+                        plugin.getGameManager().reloadConfig();
 
-                    // Reload games.
-                    plugin.getGameManager().reloadConfig();
-
-                    // Send reload messages.
-                    plugin.getMessageManager().send(sender, MessageManager.Message.RELOAD);
+                        // Send reload messages.
+                        messages.send(sender, MessageManager.Message.RELOAD);
+                    }));
                     return true;
-                case "map":
+                }
+                case "map" -> {
                     // If player doesn't have permission to get a map, send @no-permission message.
                     if (!hasPermission(player, "roulette.map")) return true;
-
                     ThreadLocalRandom random = ThreadLocalRandom.current();
                     double randomPrice = random.nextDouble(100000d);
-
                     Map.Entry<Winner.WinnerData, ItemStack> data = plugin.getWinnerManager().render(
                             player.getName(),
                             new Winner.WinnerData(
@@ -97,13 +109,15 @@ public final class Main implements CommandExecutor, TabCompleter {
                                     getRandomFromEnum(Slot.class),
                                     getRandomFromEnum(Slot.class),
                                     getRandomFromEnum(WinType.class),
-                                    randomPrice / 2));
+                                    randomPrice / 2),
+                            null);
                     if (data != null) player.getInventory().addItem(data.getValue());
-
                     return true;
-                default:
-                    plugin.getMessageManager().send(player, MessageManager.Message.SINTAX);
+                }
+                default -> {
+                    messages.send(player, MessageManager.Message.SINTAX);
                     return true;
+                }
             }
         }
 
@@ -120,18 +134,18 @@ public final class Main implements CommandExecutor, TabCompleter {
                     }
 
                     plugin.getGameManager().deleteGame(game);
-                    plugin.getMessageManager().send(player, MessageManager.Message.DELETE, message -> message.replace("%name%", game.getName()));
+                    messages.send(player, MessageManager.Message.DELETE, message -> message.replace("%name%", game.getName()));
                 } else {
-                    plugin.getMessageManager().send(player, MessageManager.Message.UNKNOWN, message -> message.replace("%name%", args[1]));
+                    messages.send(player, MessageManager.Message.UNKNOWN, message -> message.replace("%name%", args[1]));
                 }
             } else {
-                plugin.getMessageManager().send(player, MessageManager.Message.SINTAX);
+                messages.send(player, MessageManager.Message.SINTAX);
             }
             return true;
         }
 
         if (!args[0].equalsIgnoreCase("create")) {
-            plugin.getMessageManager().send(player, MessageManager.Message.SINTAX);
+            messages.send(player, MessageManager.Message.SINTAX);
             return true;
         }
 
@@ -139,7 +153,7 @@ public final class Main implements CommandExecutor, TabCompleter {
         if (!hasPermission(player, "roulette.create")) return true;
 
         if (plugin.getGameManager().exist(args[1])) {
-            plugin.getMessageManager().send(player, MessageManager.Message.EXIST, message -> message.replace("%name%", args[1]));
+            messages.send(player, MessageManager.Message.EXIST, message -> message.replace("%name%", args[1]));
             return true;
         }
 
@@ -168,34 +182,45 @@ public final class Main implements CommandExecutor, TabCompleter {
                 player.getUniqueId(),
                 ConfigManager.Config.COUNTDOWN_WAITING.asInt());
 
-        plugin.getMessageManager().send(player, MessageManager.Message.CREATE, message -> message.replace("%name%", args[1]));
-
+        messages.send(player, MessageManager.Message.CREATE, message -> message.replace("%name%", args[1]));
         return true;
     }
 
-    private <T extends Enum<T>> T getRandomFromEnum(Class<T> clazz) {
+    private <T extends Enum<T>> T getRandomFromEnum(@NotNull Class<T> clazz) {
         T[] constants = clazz.getEnumConstants();
         return constants[ThreadLocalRandom.current().nextInt(constants.length)];
     }
 
-    @Nullable
     @Override
-    public List<String> onTabComplete(@NotNull CommandSender sender, @NotNull Command command, @NotNull String alias, @NotNull String[] args) {
+    public List<String> onTabComplete(@NotNull CommandSender sender, @NotNull Command command, @NotNull String alias, @NotNull String @NotNull [] args) {
         if (args.length == 1) {
-            return StringUtil.copyPartialMatches(args[0], Arrays.asList("create", "delete", "reload", "map"), new ArrayList<>());
-        } else if (args.length == 2 && args[0].equalsIgnoreCase("delete")) {
-            List<String> games = plugin.getGameManager().getGames().stream().map(Game::getName).collect(Collectors.toList());
-            return StringUtil.copyPartialMatches(args[1], games, new ArrayList<>());
-        } else if (args.length == 3 && args[0].equalsIgnoreCase("create")) {
-            return StringUtil.copyPartialMatches(args[2], Arrays.asList("american", "european"), new ArrayList<>());
+            return StringUtil.copyPartialMatches(args[0], COMMAND_ARGS, new ArrayList<>());
         }
-        return null;
+
+        if (args.length == 2 && args[0].equalsIgnoreCase("create")) {
+            return StringUtil.copyPartialMatches(args[1], TABLE_NAME_ARG, new ArrayList<>());
+        }
+
+        if (args.length == 2 && args[0].equalsIgnoreCase("delete")) {
+            return StringUtil.copyPartialMatches(
+                    args[1],
+                    plugin.getGameManager().getGames().stream()
+                            .map(Game::getName)
+                            .collect(Collectors.toList()),
+                    new ArrayList<>());
+        }
+
+        if (args.length == 3 && args[0].equalsIgnoreCase("create")) {
+            return StringUtil.copyPartialMatches(args[2], TYPES, new ArrayList<>());
+        }
+
+        return Collections.emptyList();
     }
 
     @SuppressWarnings("BooleanMethodIsAlwaysInverted")
-    public boolean hasPermission(Player player, String permission) {
-        if (player.hasPermission(permission)) return true;
-        plugin.getMessageManager().send(player, MessageManager.Message.NOT_PERMISSION);
+    public boolean hasPermission(@NotNull CommandSender sender, String permission) {
+        if (sender.hasPermission(permission)) return true;
+        plugin.getMessageManager().send(sender, MessageManager.Message.NOT_PERMISSION);
         return false;
     }
 }
