@@ -9,13 +9,20 @@ import me.matsubara.roulette.manager.ConfigManager;
 import me.matsubara.roulette.model.stand.PacketStand;
 import me.matsubara.roulette.model.stand.StandSettings;
 import me.matsubara.roulette.util.PluginUtils;
-import org.bukkit.Bukkit;
+import org.apache.commons.lang3.ArrayUtils;
+import org.apache.commons.lang3.tuple.Pair;
+import org.bukkit.Axis;
+import org.bukkit.ChatColor;
 import org.bukkit.Location;
+import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Player;
 import org.bukkit.util.Vector;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 import java.util.Map;
+import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 @Getter
 @Setter
@@ -32,6 +39,7 @@ public final class Bet {
 
     // Slot selected.
     private Slot slot;
+    private Slot temp;
 
     // Stand showing chip.
     private PacketStand stand;
@@ -39,75 +47,103 @@ public final class Bet {
     // If this bet is in prison.
     private boolean isEnPrison;
 
-    // If this bet already was in prison.
-    private boolean wasEnPrison;
-
     // If this bet won.
     private boolean won;
 
+    // Bet offsets.
+    private Pair<Axis, double[]> offset;
+    private int offsetIndex = -1;
+
+    // The index of the glow color to use (default to WHITE).
+    private int glowColorIndex = ArrayUtils.indexOf(GLOW_COLORS, ChatColor.WHITE);
+
+    // All valid colors for glowing.
+    private static final ChatColor[] GLOW_COLORS = Stream.of(ChatColor.values())
+            .filter(ChatColor::isColor)
+            .toArray(ChatColor[]::new);
+
     public Bet(Game game) {
         this.game = game;
-        this.isEnPrison = false;
-        this.wasEnPrison = false;
-        this.won = false;
     }
 
-    /**
-     * Whether the player has a hologram.
-     */
     public boolean hasHologram() {
         return hologram != null;
     }
 
-    /**
-     * Whether the player has chosen a chip.
-     */
     public boolean hasChip() {
         return chip != null;
     }
 
-    /**
-     * Whether the player has chosen a slot.
-     */
     public boolean hasSlot() {
         return slot != null;
     }
 
-    /**
-     * Whether the player has a personal hologram for showing the chip.
-     */
     public boolean hasStand() {
         return stand != null;
     }
 
-    private void setHologram(Player player, Slot slot) {
+    public void handle(Player player, @NotNull Slot slot) {
+        this.temp = this.slot;
         this.slot = slot;
 
-        // Hologram offset, centered at the top of the chip.
-        Vector offset = PluginUtils.offsetVector(
+        this.offset = slot.getOffsets(game.getType().isEuropean());
+
+        double[] offsets = offset.getRight();
+        Axis axis = offset.getLeft();
+
+        if (temp != slot) {
+            int[] validOffsets = IntStream.range(0, offsets.length).toArray();
+
+            for (Map.Entry<Player, Bet> entry : game.getPlayers().entrySet()) {
+                Player playing = entry.getKey();
+                if (playing.equals(player)) continue;
+
+                Bet bet = entry.getValue();
+                int offset = bet.getOffsetIndex();
+
+                if (bet.getSlot() == slot && offset != -1) {
+                    validOffsets = ArrayUtils.remove(validOffsets, offset);
+                }
+
+                offsetIndex = validOffsets[0];
+            }
+
+            if (offsetIndex == -1) offsetIndex = 0;
+        }
+
+        Map.Entry<Location, StandSettings> entry = game.getModel().getLocations().get(slot.name());
+
+        Location where = entry.getKey().clone().add(PluginUtils.offsetVector(
+                new Vector(
+                        axis == Axis.X ? offsets[offsetIndex] : 0.0d,
+                        0.0d,
+                        axis == Axis.Z ? offsets[offsetIndex] : 0.0d),
+                game.getModel().getLocation().getYaw(),
+                game.getModel().getLocation().getPitch()));
+
+        setHologram(player, slot, where);
+        setStand(player, slot, where);
+    }
+
+    private void setHologram(Player player, Slot slot, @NotNull Location where) {
+        this.slot = slot;
+
+        // Where to spawn/teleport this hologram (with an offset, centered at the top of the chip).
+        Location finalWhere = where.clone().add(PluginUtils.offsetVector(
                 new Vector(
                         0.055d,
                         0.275d,
                         0.257d),
                 game.getModel().getLocation().getYaw(),
-                game.getModel().getLocation().getPitch());
-
-        // Where to spawn/teleport this hologram.
-        Location where = game
-                .getModel()
-                .getLocations()
-                .get(slot.name())
-                .getKey()
-                .clone()
-                .add(offset);
+                game.getModel().getLocation().getPitch()));
 
         // Play move chip sound at hologram location.
-        XSound.play(where, ConfigManager.Config.SOUND_SELECT.asString());
+        XSound.play(finalWhere, ConfigManager.Config.SOUND_SELECT.asString());
 
         // No need to create another hologram.
         if (!hasHologram()) {
             // Creates a personal hologram at the location of the selected slot.
-            this.hologram = new Hologram(game.getPlugin(), where);
+            this.hologram = new Hologram(game.getPlugin(), finalWhere);
             this.hologram.setVisibleByDefault(false);
             this.hologram.showTo(player);
 
@@ -118,44 +154,60 @@ public final class Bet {
                         .replace("%bet%", PluginUtils.getSlotName(slot))
                         .replace("%money%", PluginUtils.format(chip.getPrice())));
             }
-        } else {
-            hologram.teleport(where);
+            return;
+        }
 
-            // Update lines.
-            List<String> lines = ConfigManager.Config.SELECT_HOLOGRAM.asList();
-            for (int i = 0; i < lines.size(); i++) {
-                hologram.setLine(i, lines.get(i)
-                        .replace("%player%", player.getName())
-                        .replace("%bet%", PluginUtils.getSlotName(slot))
-                        .replace("%money%", PluginUtils.format(chip.getPrice())));
-            }
+        hologram.teleport(finalWhere);
+
+        // Update lines.
+        List<String> lines = ConfigManager.Config.SELECT_HOLOGRAM.asList();
+        for (int i = 0; i < lines.size(); i++) {
+            hologram.setLine(i, lines.get(i)
+                    .replace("%player%", player.getName())
+                    .replace("%bet%", PluginUtils.getSlotName(slot))
+                    .replace("%money%", PluginUtils.format(chip.getPrice())));
         }
     }
 
-    private void setStand(Slot slot) {
+    private void setStand(Player player, Slot slot, Location where) {
         this.slot = slot;
-
-        Map.Entry<Location, StandSettings> entry = game.getModel().getLocations().get(slot.name());
 
         // No need to create another stand.
         if (!hasStand()) {
             // Spawn stand with its settings.
+            StandSettings settings = game.getModel().getLocations().get(slot.name()).getValue();
+            settings.setMarker(true);
+            settings.setMainHand(PluginUtils.createHead(chip.getUrl()));
 
-            this.stand = new PacketStand(entry.getKey(), entry.getValue());
-
-            // Spawn chip item async.
-            Bukkit.getScheduler().runTaskAsynchronously(
-                    game.getPlugin(),
-                    () -> stand.setEquipment(PluginUtils.createHead(chip.getUrl()), PacketStand.ItemSlot.MAINHAND));
+            stand = new PacketStand(where, settings);
         } else {
             // Teleport.
-            this.stand.teleport(entry.getKey());
+            stand.teleport(where);
+        }
+
+        updateStandGlow(player);
+    }
+
+    public void updateStandGlow(Player player) {
+        ArmorStand standBukkit;
+        if (!hasStand() || (standBukkit = stand.getBukkitEntity()) == null) return;
+
+        try {
+            game.getPlugin().getGlowingEntities().setGlowing(
+                    stand.getEntityId(),
+                    standBukkit.getUniqueId().toString(),
+                    player,
+                    GLOW_COLORS[glowColorIndex]);
+        } catch (ReflectiveOperationException exception) {
+            exception.printStackTrace();
         }
     }
 
-    public void handle(Player player, Slot slot) {
-        setHologram(player, slot);
-        setStand(slot);
+    public void nextGlowColor() {
+        if (!hasStand()) return;
+
+        glowColorIndex++;
+        if (glowColorIndex > GLOW_COLORS.length - 1) glowColorIndex = 0;
     }
 
     public void remove() {
