@@ -10,6 +10,7 @@ import me.matsubara.roulette.game.data.Bet;
 import me.matsubara.roulette.game.data.Slot;
 import me.matsubara.roulette.gui.ChipGUI;
 import me.matsubara.roulette.gui.ConfirmGUI;
+import me.matsubara.roulette.hologram.Hologram;
 import me.matsubara.roulette.manager.ConfigManager;
 import me.matsubara.roulette.manager.MessageManager;
 import me.matsubara.roulette.model.stand.PacketStand;
@@ -22,6 +23,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.InventoryHolder;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.scheduler.BukkitRunnable;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Arrays;
 import java.util.Iterator;
@@ -30,43 +32,47 @@ import java.util.concurrent.ThreadLocalRandom;
 
 public final class Spinning extends BukkitRunnable {
 
-    // Instance not needed for now, but in the future.
-    @SuppressWarnings("all")
     private final RoulettePlugin plugin;
-
     private final Game game;
     private final PacketStand ball;
     private final boolean isEuropean;
     private final Slot[] slots;
+    private final int totalTime;
 
     private int time;
     private boolean shouldStart;
 
-    public Spinning(RoulettePlugin plugin, Game game) {
+    public Spinning(@NotNull RoulettePlugin plugin, @NotNull Game game) {
         this.plugin = plugin;
         this.game = game;
         this.ball = game.getModel().getByName("BALL");
         this.isEuropean = game.getType().isEuropean();
         this.slots = new Slot[isEuropean ? 37 : 38];
-        this.time = ConfigManager.Config.COUNTDOWN_SORTING.asInt() * 20;
+        this.totalTime = ConfigManager.Config.COUNTDOWN_SORTING.asInt() * 20;
+        this.time = totalTime;
         this.shouldStart = true;
 
         System.arraycopy(Slot.values(game), 0, slots, 0, isEuropean ? 37 : 38);
 
         game.setState(GameState.SPINNING);
 
+        Map<Player, Bet> players = game.getPlayers();
+        MessageManager messages = plugin.getMessageManager();
+
         // Check if the players selected a chip.
-        Iterator<Map.Entry<Player, Bet>> iterator = game.getPlayers().entrySet().iterator();
+        Iterator<Map.Entry<Player, Bet>> iterator = players.entrySet().iterator();
         while (iterator.hasNext()) {
+            Map.Entry<Player, Bet> entry = iterator.next();
 
             // If somehow player is null (maybe disconnected), continue.
-            Player player = iterator.next().getKey();
-            if (player == null) continue;
+            Player player = entry.getKey();
+            if (player == null || !player.isOnline()) continue;
 
             // If the player didn't select a chip, close inventory and remove from the game.
-            if (!game.getPlayers().get(player).hasChip()) {
+            Bet bet = entry.getValue();
+            if (!bet.hasChip()) {
                 game.remove(player, true);
-                plugin.getMessageManager().send(player, MessageManager.Message.OUT_OF_TIME);
+                messages.send(player, MessageManager.Message.OUT_OF_TIME);
 
                 // If the player still have the chip inventory open, close it.
                 InventoryHolder holder = player.getOpenInventory().getTopInventory().getHolder();
@@ -79,17 +85,16 @@ public final class Spinning extends BukkitRunnable {
             }
 
             // Show the bet to the player.
-            Slot selected = game.getPlayers().get(player).getSlot();
+            Slot selected = bet.getSlot();
             String numbers = selected.isDoubleZero() ? "[00]" : Arrays.toString(selected.getInts());
 
-            plugin.getMessageManager().send(player, MessageManager.Message.YOUR_BET, message -> message
+            messages.send(player, MessageManager.Message.YOUR_BET, message -> message
                     .replace("%bet%", PluginUtils.getSlotName(selected))
                     .replace("%numbers%", numbers)
                     .replace("%chance%", selected.getChance(game.getType().isEuropean())));
         }
 
-
-        if (game.getPlayers().isEmpty()) {
+        if (players.isEmpty()) {
             game.setState(GameState.ENDING);
             game.restartRunnable();
             shouldStart = false;
@@ -98,11 +103,11 @@ public final class Spinning extends BukkitRunnable {
 
         NPC npc = game.getNpc();
 
-        game.broadcast(plugin.getMessageManager().getRandomNPCMessage(npc, "no-bets"));
+        game.broadcast(messages.getRandomNPCMessage(npc, "no-bets"));
         game.broadcast(MessageManager.Message.SPINNING_START.asString());
 
         // Hide holograms to the players so everyone can see the spinning hologram.
-        game.getPlayers().forEach((player, bet) -> bet.getHologram().hideTo(player));
+        players.forEach((player, bet) -> bet.getHologram().hideTo(player));
 
         // Play NPC spin animation.
         npc.metadata().queue(PluginUtils.SNEAKING_METADATA, true).send();
@@ -120,8 +125,10 @@ public final class Spinning extends BukkitRunnable {
             return;
         }
 
+        Hologram spinHologram = game.getSpinHologram();
+
         if (time == 0) {
-            game.getSpinHologram().setLine(0, ConfigManager.Config.WINNING_NUMBER.asString());
+            spinHologram.setLine(0, ConfigManager.Config.WINNING_NUMBER.asString());
 
             // Stop NPC animation, check if there are winners and stop.
             game.getNpc().metadata().queue(PluginUtils.SNEAKING_METADATA, false).send();
@@ -134,7 +141,7 @@ public final class Spinning extends BukkitRunnable {
 
         // Spin ball.
         Location location = ball.getLocation();
-        location.setYaw(location.getYaw() + 30.0f);
+        location.setYaw(location.getYaw() + (time >= totalTime / 3 ? 30.0f : (30.0f * time / totalTime)));
         ball.teleport(location);
 
         // Select a random number.
@@ -148,23 +155,26 @@ public final class Spinning extends BukkitRunnable {
             game.setWinner(lastSpinEvent.getWinnerSlot());
         }
 
+        String slotName = PluginUtils.getSlotName(game.getWinner());
+
         // If the spin hologram is empty, create the lines, else update them.
-        if (game.getSpinHologram().size() == 0) {
+        if (spinHologram.size() == 0) {
 
             // Teleport spin hologram to its proper location.
-            game.getSpinHologram().teleport(ball
+            spinHologram.teleport(ball
                     .getLocation()
                     .clone()
                     .add(0.0d, 2.5d, 0.0d));
 
-            game.getSpinHologram().addLines(ConfigManager.Config.SPINNING.asString());
-            game.getSpinHologram().addLines(PluginUtils.getSlotName(slots[which]));
+            spinHologram.addLines(ConfigManager.Config.SPINNING.asString());
+            spinHologram.addLines(slotName);
         } else {
-            game.getSpinHologram().setLine(1, PluginUtils.getSlotName(slots[which]));
+            spinHologram.setLine(1, slotName);
 
             // Play spinning sound at spin hologram location, this sound can be heard by every player (even those outside the game).
-            XSound.play(game.getSpinHologram().getLocation(), ConfigManager.Config.SOUND_SPINNING.asString());
+            XSound.play(spinHologram.getLocation(), ConfigManager.Config.SOUND_SPINNING.asString());
         }
+
         time--;
     }
 }
