@@ -35,6 +35,7 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.BannerMeta;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.inventory.meta.SkullMeta;
+import org.jetbrains.annotations.NotNull;
 
 public final class InventoryClick implements Listener {
 
@@ -45,7 +46,7 @@ public final class InventoryClick implements Listener {
     }
 
     @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
-    public void onInventoryClick(InventoryClickEvent event) {
+    public void onInventoryClick(@NotNull InventoryClickEvent event) {
         if (event.getWhoClicked().getType() != EntityType.PLAYER) return;
 
         Player player = (Player) event.getWhoClicked();
@@ -76,7 +77,7 @@ public final class InventoryClick implements Listener {
     }
 
     @SuppressWarnings("ConstantConditions")
-    private void handleConfirmGUI(InventoryClickEvent event, Game game) {
+    private void handleConfirmGUI(@NotNull InventoryClickEvent event, Game game) {
         Player player = (Player) event.getWhoClicked();
         ItemStack item = event.getCurrentItem();
 
@@ -110,28 +111,7 @@ public final class InventoryClick implements Listener {
                     skin = "e36e94f6c34a35465fce4a90f2e25976389eb9709a12273574ff70fd4daa6852";
                 }
 
-                // Bet all chip.
-                Chip betAll = new Chip(
-                        "bet-all",
-                        skin,
-                        money);
-
-                // If the bet-all money is the same of one chip from chips.yml, use that chip.
-                for (Chip chip : plugin.getChipManager().getChips()) {
-                    if (money == chip.getPrice()) betAll = chip;
-                }
-
-                // Take money from player.
-                EconomyResponse response = plugin.getEconomy().withdrawPlayer(player, money);
-                if (!response.transactionSuccess()) {
-                    plugin.getLogger().warning(String.format("It wasn't possible to withdraw $%s to %s.", money, player.getName()));
-                    return;
-                }
-
-                plugin.getMessageManager().send(player, MessageManager.Message.SELECTED_AMOUNT, message -> message.replace("%money%", PluginUtils.format(money)));
-                plugin.getMessageManager().send(player, MessageManager.Message.CONTROL);
-
-                playerBet(game, player, betAll);
+                takeMoney(game, player, money, getBetAllChip(skin, money));
             } else if (hasDisplayName && displayName.equalsIgnoreCase(plugin.getConfigManager().getDisplayName("shop", "bet-all"))) {
                 // Do nothing.
                 return;
@@ -141,8 +121,19 @@ public final class InventoryClick implements Listener {
         closeInventory(player);
     }
 
+    private @NotNull Chip getBetAllChip(String skin, double money) {
+        Chip betAll = new Chip("bet-all", skin, money);
+
+        // If the bet-all money is the same of one chip from chips.yml, use that chip.
+        for (Chip chip : plugin.getChipManager().getChips()) {
+            if (money == chip.getPrice()) return chip;
+        }
+
+        return betAll;
+    }
+
     @SuppressWarnings("ConstantConditions")
-    private void handleChipGUI(InventoryClickEvent event, Game game) {
+    private void handleChipGUI(@NotNull InventoryClickEvent event, Game game) {
         Player player = (Player) event.getWhoClicked();
         ItemStack item = event.getCurrentItem();
         ChipGUI holder = (ChipGUI) event.getClickedInventory().getHolder();
@@ -160,7 +151,7 @@ public final class InventoryClick implements Listener {
             // Open confirm gui.
             int current = holder.getCurrent();
             runTask(() -> {
-                ConfirmGUI gui = new ConfirmGUI(plugin, player, ConfirmGUI.ConfirmType.BET_ALL);
+                ConfirmGUI gui = new ConfirmGUI(game, player, ConfirmGUI.ConfirmType.BET_ALL);
                 gui.setPreviousPage(current);
             });
             return;
@@ -194,25 +185,32 @@ public final class InventoryClick implements Listener {
                 return;
             }
 
-            // Take money from player.
-            EconomyResponse response = plugin.getEconomy().withdrawPlayer(player, money);
-            if (!response.transactionSuccess()) {
-                plugin.getLogger().warning(String.format("It wasn't possible to withdraw $%s to %s.", money, player.getName()));
-                return;
-            }
-
-            // Remove the money and close inventory.
-            plugin.getMessageManager().send(player, MessageManager.Message.SELECTED_AMOUNT, message -> message.replace("%money%", PluginUtils.format(money)));
-            plugin.getMessageManager().send(player, MessageManager.Message.CONTROL);
-
-            playerBet(game, player, chip);
+            takeMoney(game, player, money, chip);
         } else return;
 
         closeInventory(player);
     }
 
+    private void takeMoney(Game game, Player player, double money, Chip chip) {
+        // Take money from player.
+        EconomyResponse response = plugin.getEconomy().withdrawPlayer(player, money);
+        if (!response.transactionSuccess()) {
+            plugin.getLogger().warning(String.format("It wasn't possible to withdraw $%s to %s.", money, player.getName()));
+            return;
+        }
+
+        // Remove the money and close inventory.
+        MessageManager messages = plugin.getMessageManager();
+        messages.send(player, MessageManager.Message.SELECTED_AMOUNT, message -> message
+                .replace("%money%", PluginUtils.format(money))
+                .replace("%money-left%", PluginUtils.format(plugin.getEconomy().getBalance(player))));
+        messages.send(player, MessageManager.Message.CONTROL);
+
+        playerBet(game, player, chip);
+    }
+
     @SuppressWarnings("ConstantConditions")
-    private void handleGameGUI(InventoryClickEvent event) {
+    private void handleGameGUI(@NotNull InventoryClickEvent event) {
         GameGUI holder = (GameGUI) event.getClickedInventory().getHolder();
         Game game = holder.getGame();
 
@@ -225,33 +223,37 @@ public final class InventoryClick implements Listener {
         boolean hasDisplayName = current.getItemMeta().hasDisplayName();
         String displayName = current.getItemMeta().getDisplayName();
 
-        String noAccount = plugin.getConfigManager().getDisplayName("game-menu", "no-account");
-        String minAmount = plugin.getConfigManager().getDisplayName("game-menu", "min-amount");
-        String maxAmount = plugin.getConfigManager().getDisplayName("game-menu", "max-amount");
-        String startTime = plugin.getConfigManager().getStartTimeDisplayName(game.getStartTime());
-        String laPartage = plugin.getConfigManager().getDisplayName("game-menu", "la-partage");
-        String enPrison = plugin.getConfigManager().getDisplayName("game-menu", "en-prison");
-        String surrender = plugin.getConfigManager().getDisplayName("game-menu", "surrender");
+        ConfigManager configManager = plugin.getConfigManager();
+
+        String noAccount = configManager.getDisplayName("game-menu", "no-account");
+        String minAmount = configManager.getDisplayName("game-menu", "min-amount");
+        String maxAmount = configManager.getDisplayName("game-menu", "max-amount");
+        String startTime = configManager.getStartTimeDisplayName(game.getStartTime());
+        String laPartage = configManager.getDisplayName("game-menu", "la-partage");
+        String enPrison = configManager.getDisplayName("game-menu", "en-prison");
+        String surrender = configManager.getDisplayName("game-menu", "surrender");
 
         String state = holder.getGame().isBetAll() ? ConfigManager.Config.STATE_ENABLED.asString() : ConfigManager.Config.STATE_DISABLED.asString();
-        String betAll = plugin.getConfigManager().getDisplayName("game-menu", "bet-all").replace("%state%", state);
+        String betAll = configManager.getDisplayName("game-menu", "bet-all").replace("%state%", state);
 
-        String close = plugin.getConfigManager().getDisplayName("game-menu", "close");
+        String close = configManager.getDisplayName("game-menu", "close");
 
         String npcName = game.getNPCName();
         if (npcName == null) npcName = ConfigManager.Config.UNNAMED_CROUPIER.asString();
-        String croupier = plugin.getConfigManager().getDisplayName("game-menu", "croupier").replace("%croupier-name%", npcName);
+        String croupier = configManager.getDisplayName("game-menu", "croupier").replace("%croupier-name%", npcName);
+
+        MessageManager messages = plugin.getMessageManager();
 
         if (hasDisplayName && (displayName.equalsIgnoreCase(minAmount) || displayName.equalsIgnoreCase(maxAmount))) {
             setLimitPlayers(event, game, min, max, displayName.equalsIgnoreCase(maxAmount));
         } else if (isAccountItem(current) || (hasDisplayName && displayName.equalsIgnoreCase(noAccount))) {
             if (current.getType() == XMaterial.PLAYER_HEAD.parseMaterial() && event.getClick() == ClickType.RIGHT) {
                 game.setAccountGiveTo(null);
-                plugin.getMessageManager().send(player, MessageManager.Message.NO_ACCOUNT);
-                event.setCurrentItem(plugin.getConfigManager().getItem("game-menu", "no-account", null));
+                messages.send(player, MessageManager.Message.NO_ACCOUNT);
+                event.setCurrentItem(configManager.getItem("game-menu", "no-account", null));
             } else {
                 plugin.getInputManager().newInput(player, InputManager.InputType.ACCOUNT_NAME, game);
-                plugin.getMessageManager().send(player, MessageManager.Message.NPC_NAME);
+                messages.send(player, MessageManager.Message.NPC_NAME);
             }
             closeInventory(player);
         } else if (hasDisplayName && displayName.equalsIgnoreCase(startTime)) {
@@ -261,14 +263,14 @@ public final class InventoryClick implements Listener {
             // Can't be null since the names are equals to the rules ones.
             GameRule rule = GameRule.valueOf(NBTEditor.getString(current, "rouletteRule"));
             if (rule.isSurrender() && !game.getType().isAmerican()) {
-                plugin.getMessageManager().send(player, MessageManager.Message.ONLY_AMERICAN);
+                messages.send(player, MessageManager.Message.ONLY_AMERICAN);
                 closeInventory(player);
                 return;
             }
 
             // Prison rule can only be applied in a game with 1 min player required.
             if (rule.isEnPrison() && game.getMinPlayers() > 1) {
-                plugin.getMessageManager().send(player, MessageManager.Message.PRISON_ERROR);
+                messages.send(player, MessageManager.Message.PRISON_ERROR);
                 closeInventory(player);
                 return;
             }
@@ -291,21 +293,21 @@ public final class InventoryClick implements Listener {
         } else if (hasDisplayName && ChatColor.stripColor(displayName).equalsIgnoreCase(ChatColor.stripColor(croupier))) {
             if (event.getClick() == ClickType.LEFT) {
                 plugin.getInputManager().newInput(player, InputManager.InputType.CROUPIER_NAME, game);
-                plugin.getMessageManager().send(player, MessageManager.Message.NPC_NAME);
+                messages.send(player, MessageManager.Message.NPC_NAME);
             } else if (event.getClick() == ClickType.RIGHT) {
                 plugin.getInputManager().newInput(player, InputManager.InputType.CROUPIER_TEXTURE, game);
-                plugin.getMessageManager().send(player, MessageManager.Message.NPC_TEXTURE);
+                messages.send(player, MessageManager.Message.NPC_TEXTURE);
             } else if (event.getClick() == ClickType.SHIFT_LEFT) {
                 String texture = game.getNPCTexture();
                 String signature = game.getNPCSignature();
 
-                plugin.getMessageManager().send(player, MessageManager.Message.NPC_RENAMED);
+                messages.send(player, MessageManager.Message.NPC_RENAMED);
                 game.setNPC(null, texture, signature);
                 plugin.getGameManager().save(game);
             } else if (event.getClick() == ClickType.SHIFT_RIGHT) {
                 String name = game.getNPCName() == null ? "" : game.getNPCName();
 
-                plugin.getMessageManager().send(player, MessageManager.Message.NPC_TEXTURIZED);
+                messages.send(player, MessageManager.Message.NPC_TEXTURIZED);
                 game.setNPC(name, null, null);
                 plugin.getGameManager().save(game);
             }
@@ -321,7 +323,7 @@ public final class InventoryClick implements Listener {
     }
 
     @SuppressWarnings("ConstantConditions")
-    private void disableRule(Game game, Inventory inventory, GameRule... rules) {
+    private void disableRule(Game game, Inventory inventory, GameRule @NotNull ... rules) {
         for (GameRule rule : rules) {
             if (!game.isRuleEnabled(rule)) continue;
 
@@ -330,22 +332,22 @@ public final class InventoryClick implements Listener {
         }
     }
 
-    private void setBannerColor(ItemStack item, boolean enabled) {
+    private void setBannerColor(@NotNull ItemStack item, boolean enabled) {
         BannerMeta meta = (BannerMeta) item.getItemMeta();
-        if (meta != null) {
-            meta.addPattern(new Pattern(enabled ? DyeColor.LIME : DyeColor.RED, PatternType.BASE));
-            item.setItemMeta(meta);
-        }
+        if (meta == null) return;
+
+        meta.addPattern(new Pattern(enabled ? DyeColor.LIME : DyeColor.RED, PatternType.BASE));
+        item.setItemMeta(meta);
     }
 
 
-    private boolean isAccountItem(ItemStack item) {
+    private boolean isAccountItem(@NotNull ItemStack item) {
         //noinspection ConstantConditions, already checked in the event.
         return item.getType() == XMaterial.PLAYER_HEAD.parseMaterial() && ((SkullMeta) item.getItemMeta()).hasOwner();
     }
 
     @SuppressWarnings("ConstantConditions")
-    private void setLimitPlayers(InventoryClickEvent event, Game game, int min, int max, boolean isMax) {
+    private void setLimitPlayers(@NotNull InventoryClickEvent event, Game game, int min, int max, boolean isMax) {
         if (event.getClick() == ClickType.LEFT) {
             game.setLimitPlayers(!isMax ? min - 1 : min, !isMax ? max : max - 1);
         } else if (event.getClick() == ClickType.RIGHT) {
@@ -368,7 +370,7 @@ public final class InventoryClick implements Listener {
     }
 
     @SuppressWarnings("ConstantConditions")
-    private void setStartTime(InventoryClickEvent event, Game game) {
+    private void setStartTime(@NotNull InventoryClickEvent event, @NotNull Game game) {
         int time = game.getStartTime();
         if (event.getClick() == ClickType.LEFT) {
             time -= 5;
@@ -380,21 +382,23 @@ public final class InventoryClick implements Listener {
 
         game.setStartTime(time);
 
-        if (event.getCurrentItem().getAmount() != time) {
-            event.getCurrentItem().setAmount(time);
+        ItemStack item = event.getCurrentItem();
 
-            ItemMeta meta = event.getCurrentItem().getItemMeta();
+        if (item.getAmount() != time) {
+            item.setAmount(time);
+
+            ItemMeta meta = item.getItemMeta();
             if (meta == null) return;
 
             meta.setDisplayName(plugin.getConfigManager().getStartTimeDisplayName(time));
-            event.getCurrentItem().setItemMeta(meta);
+            item.setItemMeta(meta);
         }
 
         // Save data.
         plugin.getGameManager().save(game);
     }
 
-    private void closeInventory(Player player) {
+    private void closeInventory(@NotNull Player player) {
         runTask(player::closeInventory);
     }
 
@@ -402,12 +406,15 @@ public final class InventoryClick implements Listener {
         plugin.getServer().getScheduler().runTask(plugin, runnable);
     }
 
-    private void playerBet(Game game, Player player, Chip chip) {
+    private void playerBet(@NotNull Game game, Player player, Chip chip) {
         // Save player chip.
         game.getPlayers().get(player).setChip(chip);
 
         // First automated move.
         game.moveChip(player, true);
+
+        // Send glow advice message.
+        plugin.getMessageManager().send(player, MessageManager.Message.CHANGE_GLOW_COLOR);
 
         // If the money animation isn't running, run now.
         if (game.getMoneyAnimation() == null) {
@@ -418,26 +425,28 @@ public final class InventoryClick implements Listener {
     }
 
     @SuppressWarnings("ConstantConditions")
-    private void setBetAll(InventoryClickEvent event, Game game) {
+    private void setBetAll(@NotNull InventoryClickEvent event, @NotNull Game game) {
         boolean betAll = !game.isBetAll();
 
         game.setBetAll(betAll);
 
         String state = game.isBetAll() ? ConfigManager.Config.STATE_ENABLED.asString() : ConfigManager.Config.STATE_DISABLED.asString();
 
-        ItemMeta meta = event.getCurrentItem().getItemMeta();
+        ItemStack item = event.getCurrentItem();
+        ItemMeta meta = item.getItemMeta();
+
         meta.setDisplayName(plugin.getConfigManager().getDisplayName("game-menu", "bet-all").replace("%state%", state));
         if (!meta.hasItemFlag(ItemFlag.HIDE_ATTRIBUTES)) meta.addItemFlags(ItemFlag.HIDE_ENCHANTS);
 
         if (betAll) {
-            event.getCurrentItem().addUnsafeEnchantment(Enchantment.ARROW_DAMAGE, 1);
+            item.addUnsafeEnchantment(Enchantment.ARROW_DAMAGE, 1);
         } else {
-            if (event.getCurrentItem().containsEnchantment(Enchantment.ARROW_DAMAGE)) {
-                event.getCurrentItem().removeEnchantment(Enchantment.ARROW_DAMAGE);
+            if (item.containsEnchantment(Enchantment.ARROW_DAMAGE)) {
+                item.removeEnchantment(Enchantment.ARROW_DAMAGE);
             }
         }
 
-        event.getCurrentItem().setItemMeta(meta);
+        item.setItemMeta(meta);
 
         // Save data.
         plugin.getGameManager().save(game);
