@@ -9,10 +9,12 @@ import me.matsubara.roulette.RoulettePlugin;
 import me.matsubara.roulette.game.Game;
 import me.matsubara.roulette.game.GameRule;
 import me.matsubara.roulette.game.GameState;
+import me.matsubara.roulette.game.data.Bet;
 import me.matsubara.roulette.gui.ConfirmGUI;
 import me.matsubara.roulette.manager.ConfigManager;
 import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -24,13 +26,15 @@ public final class SteerVehicle extends PacketAdapter {
     private final Map<UUID, Long> steerCooldown;
 
     public SteerVehicle(RoulettePlugin plugin) {
-        super(plugin, ListenerPriority.HIGHEST, PacketType.Play.Client.STEER_VEHICLE);
+        super(PacketAdapter.params(plugin, PacketType.Play.Client.STEER_VEHICLE)
+                .listenerPriority(ListenerPriority.HIGHEST)
+                .optionSync());
         this.plugin = plugin;
         this.steerCooldown = new HashMap<>();
     }
 
     @Override
-    public void onPacketReceiving(PacketEvent event) {
+    public void onPacketReceiving(@NotNull PacketEvent event) {
         Player player = event.getPlayer();
 
         // Get required values from the packet.
@@ -38,10 +42,10 @@ public final class SteerVehicle extends PacketAdapter {
         float sideways = floats.readSafely(0);
 
         StructureModifier<Boolean> booleans = event.getPacket().getBooleans();
-        boolean dismount = booleans.readSafely(1);
+        boolean jump = booleans.readSafely(0), dismount = booleans.readSafely(1);
 
-        // At the moment, we only care about left/right and shift keys.
-        if (sideways == 0.0f && !dismount) return;
+        // At the moment, we only care about left/right, space and shift keys.
+        if (sideways == 0.0f && !jump && !dismount) return;
 
         // Let player dismount if is a passenger of a packet stand.
         models:
@@ -62,53 +66,59 @@ public final class SteerVehicle extends PacketAdapter {
                     return;
                 }
 
-                // Run sync to prevent exceptions.
-                plugin.getServer().getScheduler().runTask(plugin, () -> {
-                    // Left
-                    if (sideways > 0.0f) {
-                        if ((state.isIdle() || state.isStarting()) && canSwapChair(player, game)) {
-                            game.sitPlayer(player, false);
-                        } else {
-                            // Only allow move chip if the bet isn't in prison.
-                            if (!game.isRuleEnabled(GameRule.EN_PRISON) || !game.getPlayers().get(player).isEnPrison()) {
-                                game.moveChip(player, false);
-                            }
+                Bet bet = game.getPlayers().get(player);
+
+                // Left.
+                if (sideways > 0.0f) {
+                    if ((state.isIdle() || state.isStarting()) && canSwapChair(player, game)) {
+                        game.sitPlayer(player, false);
+                    } else {
+                        // Only allow move chip if the bet isn't in prison.
+                        if (!game.isRuleEnabled(GameRule.EN_PRISON) || !bet.isEnPrison()) {
+                            game.moveChip(player, false);
                         }
                     }
+                }
 
-                    // Right
-                    if (sideways < 0.0f) {
-                        if ((state.isIdle() || state.isStarting()) && canSwapChair(player, game)) {
-                            game.sitPlayer(player, true);
-                        } else {
-                            // Only allow move chip if the bet isn't in prison.
-                            if (!game.isRuleEnabled(GameRule.EN_PRISON) || !game.getPlayers().get(player).isEnPrison()) {
-                                game.moveChip(player, true);
-                            }
+                // Right.
+                if (sideways < 0.0f) {
+                    if ((state.isIdle() || state.isStarting()) && canSwapChair(player, game)) {
+                        game.sitPlayer(player, true);
+                    } else {
+                        // Only allow move chip if the bet isn't in prison.
+                        if (!game.isRuleEnabled(GameRule.EN_PRISON) || !bet.isEnPrison()) {
+                            game.moveChip(player, true);
                         }
                     }
+                }
 
-                    // Shift
-                    if (dismount) {
-                        if (game.getState().isEnding() && !game.isRuleEnabled(GameRule.EN_PRISON)) {
-                            // Let player unmount.
-                            game.kickPlayer(player);
-                        } else {
-                            // Open leave confirm gui to the player sync.
-                            new ConfirmGUI(plugin, player, ConfirmGUI.ConfirmType.LEAVE);
-                        }
+                // Shift.
+                if (dismount) {
+                    if (state.isEnding() && !game.isRuleEnabled(GameRule.EN_PRISON)) {
+                        // Let player unmount.
+                        game.kickPlayer(player);
+                    } else {
+                        // Open leave confirm gui to the player sync.
+                        new ConfirmGUI(game, player, ConfirmGUI.ConfirmType.LEAVE);
                     }
+                }
 
-                    // Put player in cooldown.
-                    steerCooldown.put(player.getUniqueId(), System.currentTimeMillis() + ConfigManager.Config.MOVE_INTERVAL.asLong());
-                });
+                // Jump.
+                if (jump && (state.isSelecting() || state.isSpinning())) {
+                    bet.nextGlowColor();
+                    bet.updateStandGlow(player);
+                }
+
+                // Put player in cooldown.
+                long interval = Math.max(200L, ConfigManager.Config.MOVE_INTERVAL.asLong());
+                steerCooldown.put(player.getUniqueId(), System.currentTimeMillis() + interval);
 
                 break models;
             }
         }
     }
 
-    private boolean canSwapChair(Player player, Game game) {
-        return (game.getState().isIdle() || game.getState().isStarting()) && (ConfigManager.Config.SWAP_CHAIR.asBoolean() || player.hasPermission("roulette.swap_chair"));
+    private boolean canSwapChair(Player player, @NotNull Game game) {
+        return (game.getState().isIdle() || game.getState().isStarting()) && (ConfigManager.Config.SWAP_CHAIR.asBool() || player.hasPermission("roulette.swapchair"));
     }
 }
