@@ -3,6 +3,7 @@ package me.matsubara.roulette.listener.protocol;
 import com.comphenix.protocol.PacketType;
 import com.comphenix.protocol.events.ListenerPriority;
 import com.comphenix.protocol.events.PacketAdapter;
+import com.comphenix.protocol.events.PacketContainer;
 import com.comphenix.protocol.events.PacketEvent;
 import com.comphenix.protocol.wrappers.EnumWrappers;
 import com.cryptomorin.xseries.ReflectionUtils;
@@ -13,58 +14,65 @@ import me.matsubara.roulette.RoulettePlugin;
 import me.matsubara.roulette.event.PlayerRouletteEnterEvent;
 import me.matsubara.roulette.game.Game;
 import me.matsubara.roulette.manager.MessageManager;
+import me.matsubara.roulette.model.Model;
 import me.matsubara.roulette.model.stand.PacketStand;
-import me.matsubara.roulette.util.Lang3Utils;
+import org.apache.commons.lang3.ArrayUtils;
+import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.metadata.MetadataValue;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
-import java.util.Collection;
 import java.util.Iterator;
-import java.util.stream.IntStream;
+import java.util.Map;
 
 public final class UseEntity extends PacketAdapter {
 
     private final RoulettePlugin plugin;
 
     public UseEntity(RoulettePlugin plugin) {
-        super(plugin, ListenerPriority.HIGHEST, PacketType.Play.Client.USE_ENTITY);
+        super(PacketAdapter.params(plugin, PacketType.Play.Client.USE_ENTITY)
+                .listenerPriority(ListenerPriority.HIGHEST)
+                .optionSync());
         this.plugin = plugin;
     }
 
     @Override
-    public void onPacketReceiving(PacketEvent event) {
-        int entityId = event.getPacket().getIntegers().readSafely(0);
+    public void onPacketReceiving(@NotNull PacketEvent event) {
+        PacketContainer packet = event.getPacket();
+        int entityId = packet.getIntegers().readSafely(0);
 
         EnumWrappers.EntityUseAction action;
         if (ReflectionUtils.MINOR_NUMBER > 16) {
-            action = event.getPacket().getEnumEntityUseActions().readSafely(0).getAction();
+            action = packet.getEnumEntityUseActions().readSafely(0).getAction();
         } else {
-            action = event.getPacket().getEntityUseActions().readSafely(0);
+            action = packet.getEntityUseActions().readSafely(0);
         }
 
-        boolean isLeft = action == EnumWrappers.EntityUseAction.ATTACK;
-        if (isLeft) return;
+        if (action == EnumWrappers.EntityUseAction.ATTACK) return;
 
         Player player = event.getPlayer();
 
-        plugin.getServer().getScheduler().runTask(plugin, () -> {
-            for (Game game : plugin.getGameManager().getGames()) {
-                handleInteract(game, player, entityId, game.getModel().getStands().values(), game.getJoinHologram().getStands());
-            }
-        });
+        for (Game game : plugin.getGameManager().getGames()) {
+            handleInteract(game, player, entityId, game.getModel().getStands(), game.getJoinHologram().getStands());
+        }
     }
 
     @SuppressWarnings("OptionalGetWithoutIsPresent")
     @SafeVarargs
-    private final void handleInteract(Game game, Player player, int entityId, Collection<PacketStand>... collections) {
-        for (Collection<PacketStand> stands : collections) {
-            for (PacketStand stand : stands) {
+    private void handleInteract(Game game, Player player, int entityId, Map<String, PacketStand> @NotNull ... collections) {
+        MessageManager messages = plugin.getMessageManager();
+        for (Map<String, PacketStand> collection : collections) {
+            for (Map.Entry<String, PacketStand> entry : collection.entrySet()) {
+                PacketStand stand = entry.getValue();
                 if (stand == null || stand.getEntityId() != entityId) continue;
 
+                Model model = game.getModel();
+
                 // Can happen when the game is created.
-                if (!game.getModel().isModelSpawned()) {
-                    plugin.getMessageManager().send(player, MessageManager.Message.MODEL_NOT_LOADED);
+                if (!model.isModelSpawned()) {
+                    messages.send(player, MessageManager.Message.MODEL_NOT_LOADED);
                     return;
                 }
 
@@ -75,24 +83,24 @@ public final class UseEntity extends PacketAdapter {
                     boolean isPlanks = material.name().contains("PLANKS");
 
                     // No need to change planks type if is the same.
-                    if (isPlanks && material == game.getModel().getPlanksType()) return;
+                    if (isPlanks && material == model.getPlanksType()) return;
 
                     String toUse = material.name().substring(0, material.name().lastIndexOf("_"));
 
                     if (isPlanks) {
                         XMaterial slab = XMaterial.matchXMaterial(toUse + "_SLAB").get(); // Can't be null.
-                        game.getModel().setPlanksType(material);
-                        game.getModel().setSlabsType(slab);
+                        model.setPlanksType(material);
+                        model.setSlabsType(slab);
                     } else {
                         XMaterial carpet = XMaterial.matchXMaterial(toUse + "_CARPET").get(); // Again, can't be null.
-                        game.getModel().setCarpetsType(carpet);
+                        model.setCarpetsType(carpet);
                     }
 
-                    ItemStack carpets = game.getModel().getCarpetsType().parseItem();
-                    ItemStack planks = game.getModel().getPlanksType().parseItem();
-                    ItemStack slabs = game.getModel().getSlabsType().parseItem();
+                    ItemStack carpets = model.getCarpetsType().parseItem();
+                    ItemStack planks = model.getPlanksType().parseItem();
+                    ItemStack slabs = model.getSlabsType().parseItem();
 
-                    game.getModel().getStands().forEach((name, part) -> {
+                    model.getStands().forEach((name, part) -> {
                         if (isPlanks) {
                             if (name.startsWith("SIDE")) {
                                 part.setEquipment(slabs, PacketStand.ItemSlot.HEAD);
@@ -105,9 +113,9 @@ public final class UseEntity extends PacketAdapter {
 
                         if (name.startsWith("CHAIR")) {
                             int current = Integer.parseInt(name.split("_")[1]);
-                            if (Lang3Utils.contains(IntStream.range(0, 10).map(x -> (x * 3) + 1).toArray(), current)) {
+                            if (ArrayUtils.contains(Model.CHAIR_FIRST_LAYER, current)) {
                                 part.setEquipment(planks, PacketStand.ItemSlot.HEAD);
-                            } else if (Lang3Utils.contains(IntStream.range(0, 10).map(x -> (x * 3) + 2).toArray(), current)) {
+                            } else if (ArrayUtils.contains(Model.CHAIR_SECOND_LAYER, current)) {
                                 part.setEquipment(slabs, PacketStand.ItemSlot.HEAD);
                             } else {
                                 part.setEquipment(carpets, PacketStand.ItemSlot.HEAD);
@@ -119,38 +127,41 @@ public final class UseEntity extends PacketAdapter {
                     return;
                 }
 
+                // Joining while shifting will remove the player from the chair.
+                if (player.isSneaking()) return;
+
                 // Can happen when the player is already in the game.
                 if (game.isPlaying(player) && game.isSittingOn(player)) {
-                    plugin.getMessageManager().send(player, MessageManager.Message.ALREADY_INGAME);
+                    messages.send(player, MessageManager.Message.ALREADY_INGAME);
                     return;
                 }
 
                 // Can happen whe the game is already started.
                 if (!game.canJoin()) {
                     if (game.getState().isEnding()) {
-                        plugin.getMessageManager().send(player, MessageManager.Message.RESTARTING);
+                        messages.send(player, MessageManager.Message.RESTARTING);
                     } else {
-                        plugin.getMessageManager().send(player, MessageManager.Message.ALREADY_STARTED);
+                        messages.send(player, MessageManager.Message.ALREADY_STARTED);
                     }
                     return;
                 }
 
                 // Can happen when the game is full.
                 if (game.isFull()) {
-                    plugin.getMessageManager().send(player, MessageManager.Message.FULL);
+                    messages.send(player, MessageManager.Message.FULL);
                     return;
                 }
 
                 // Check if player is vanished using EssentialsX, SuperVanish, PremiumVanish, VanishNoPacket, etc.
                 if (isPluginVanished(player)) {
-                    plugin.getMessageManager().send(player, MessageManager.Message.VANISH);
+                    messages.send(player, MessageManager.Message.VANISH);
                     return;
                 }
 
                 // Can happen if the player doens't have money.
                 double minAmount = plugin.getChipManager().getMinAmount();
                 if (!plugin.getEconomy().has(player, minAmount)) {
-                    plugin.getMessageManager().send(player, MessageManager.Message.MIN_REQUIRED, message -> message.replace("%money%", String.valueOf(minAmount)));
+                    messages.send(player, MessageManager.Message.MIN_REQUIRED, message -> message.replace("%money%", String.valueOf(minAmount)));
                     return;
                 }
 
@@ -158,7 +169,10 @@ public final class UseEntity extends PacketAdapter {
                 plugin.getServer().getPluginManager().callEvent(enterEvent);
                 if (enterEvent.isCancelled()) return;
 
-                game.add(player);
+                Integer sitAt = handleChairInteract(game, player, entry.getKey());
+                if (sitAt == null) return;
+
+                game.add(player, sitAt);
                 game.broadcast(MessageManager.Message.JOIN.asString()
                         .replace("%player%", player.getName())
                         .replace("%playing%", String.valueOf(game.size()))
@@ -168,7 +182,24 @@ public final class UseEntity extends PacketAdapter {
         }
     }
 
-    private XMaterial getMaterialInHand(Player player) {
+    private @Nullable Integer handleChairInteract(Game game, Player player, @NotNull String name) {
+        if (!name.startsWith("CHAIR")) return -1;
+
+        int which = Integer.parseInt(name.split("_")[1]);
+        if (!ArrayUtils.contains(Model.CHAIR_SECOND_LAYER, which)) return -1;
+
+        ArmorStand chair = game.getChair(which);
+        if (chair == null) return -1;
+
+        if (!chair.getPassengers().isEmpty()) {
+            plugin.getMessageManager().send(player, MessageManager.Message.SEAT_TAKEN);
+            return null;
+        }
+
+        return which;
+    }
+
+    private XMaterial getMaterialInHand(@NotNull Player player) {
         XMaterial material;
         try {
             material = XMaterial.matchXMaterial(player.getInventory().getItemInMainHand());
@@ -181,10 +212,10 @@ public final class UseEntity extends PacketAdapter {
     private boolean canChangeTexture(Player player) {
         String inHand = getMaterialInHand(player).name();
         boolean validMaterial = inHand.contains("PLANKS") || inHand.contains("WOOL");
-        return player.isSneaking() && validMaterial && player.hasPermission("roulette.texture_table");
+        return player.isSneaking() && validMaterial && player.hasPermission("roulette.texturetable");
     }
 
-    private boolean isPluginVanished(Player player) {
+    private boolean isPluginVanished(@NotNull Player player) {
         Iterator<MetadataValue> iterator = player.getMetadata("vanished").iterator();
 
         MetadataValue meta;
