@@ -7,16 +7,18 @@ import me.matsubara.roulette.util.PluginUtils;
 import org.bukkit.Material;
 import org.bukkit.inventory.ItemFlag;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.plugin.java.JavaPlugin;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 
 public final class ConfigManager {
 
-    private static RoulettePlugin plugin;
+    private final RoulettePlugin plugin;
 
     public ConfigManager(RoulettePlugin plugin) {
-        ConfigManager.plugin = plugin;
+        this.plugin = plugin;
     }
 
     public int getRenderDistance() {
@@ -35,27 +37,27 @@ public final class ConfigManager {
         return PluginUtils.translate(plugin.getConfig().getString("slots." + type + "." + index));
     }
 
-    public String getAccountDisplayName(String name) {
+    public @NotNull String getAccountDisplayName(String name) {
         return getDisplayName("game-menu", "account").replace("%player%", name);
     }
 
-    public List<String> getAccountLore() {
+    public @NotNull List<String> getAccountLore() {
         return getLore("game-menu", "account");
     }
 
-    public String getStartTimeDisplayName(int time) {
+    public @NotNull String getStartTimeDisplayName(int time) {
         return getDisplayName("game-menu", "start-time").replace("%seconds%", String.valueOf(time));
     }
 
-    public String getChipDisplayName(double price) {
+    public @NotNull String getChipDisplayName(double price) {
         return getDisplayName("shop", "chip").replace("%money%", PluginUtils.format(price));
     }
 
-    public List<String> getChipLore() {
+    public @NotNull List<String> getChipLore() {
         return getLore("shop", "chip");
     }
 
-    public ItemStack getItem(String gui, String type, @Nullable String money) {
+    public @Nullable ItemStack getItem(String gui, String type, @Nullable String money) {
         Material material = XMaterial.matchXMaterial(getMaterial(gui, type)).map(XMaterial::parseMaterial).orElse(null);
         if (material == null) return null;
 
@@ -79,11 +81,11 @@ public final class ConfigManager {
         return plugin.getConfig().contains(gui + "." + path + ".attributes", false);
     }
 
-    private ItemFlag[] getAttributes(String gui, String path) {
+    private ItemFlag @NotNull [] getAttributes(String gui, String path) {
         return plugin.getConfig().getStringList(gui + "." + path + ".attributes").stream().map(this::getAttribute).toArray(ItemFlag[]::new);
     }
 
-    private ItemFlag getAttribute(String attribute) {
+    private @Nullable ItemFlag getAttribute(String attribute) {
         try {
             return ItemFlag.valueOf(attribute);
         } catch (IllegalArgumentException exception) {
@@ -108,15 +110,16 @@ public final class ConfigManager {
         return PluginUtils.translate(plugin.getConfig().getString(gui + "." + path + ".display-name"));
     }
 
-    public List<String> getLore(String gui, String path) {
+    public @NotNull List<String> getLore(String gui, String path) {
         return PluginUtils.translate(plugin.getConfig().getStringList(gui + "." + path + ".lore"));
     }
 
     public enum Config {
-        SWAP_CHAIR("swap-schair"),
+        SWAP_CHAIR("swap-chair"),
         INSTANT_EXPLODE("instant-explode"),
         FIX_CHAIR_CAMERA("fix-chair-camera"),
         HIT_ON_GAME("hit-on-game"),
+        KEEP_SEAT("keep-seat"),
         LEAVE_CONFIRM("leave-confirm"),
         MOVE_INTERVAL("move-interval"),
         CROUPIER_BALL("croupier-ball.material"),
@@ -126,15 +129,17 @@ public final class ConfigManager {
         COUNTDOWN_SORTING("countdown.sorting"),
         RESTART_TIME("restart.time"),
         RESTART_FIREWORKS("restart.fireworks"),
-        SOUND_CLICK("sound.click"),
-        SOUND_COUNTDOWN("sound.countdown"),
-        SOUND_SPINNING("sound.spinning"),
-        SOUND_SWAP_CHAIR("sound.swap-chair"),
-        SOUND_SELECT("sound.select"),
+        SOUND_CLICK("sounds.click"),
+        SOUND_COUNTDOWN("sounds.countdown"),
+        SOUND_SPINNING("sounds.spinning"),
+        SOUND_SWAP_CHAIR("sounds.swap-chair"),
+        SOUND_SELECT("sounds.select"),
         DISABLED_SLOTS("disabled-slots"),
         MAP_IMAGE_ENABLED("map-image.enabled"),
-        MAP_IMAGE_SHIFT_ID_FIX("map-image.shift-id-fix"),
+        MAP_IMAGE_DATE_FORMAT("map-image.date-format"),
         MAP_IMAGE_TEXT("map-image.text"),
+        MAP_IMAGE_ITEM_DISPLAY_NAME("map-image.item.display-name"),
+        MAP_IMAGE_ITEM_LORE("map-image.item.lore"),
         CANCEL_WORD("cancel-word"),
         SPINNING("spin-holograms.spinning"),
         WINNING_NUMBER("spin-holograms.winning-number"),
@@ -147,8 +152,8 @@ public final class ConfigManager {
         ODD("slots.other.odd"),
         RED("slots.other.red"),
         BLACK("slots.other.black"),
-        TYPE_EUROPEAN("type.european"),
-        TYPE_AMERICAN("type.american"),
+        TYPE_EUROPEAN("types.european"),
+        TYPE_AMERICAN("types.american"),
         CONFIRM_GUI_TITLE("confirmation-gui.title"),
         CONFIRM_GUI_CONFIRM("confirmation-gui.confirm"),
         CONFIRM_GUI_CANCEL("confirmation-gui.cancel"),
@@ -166,6 +171,7 @@ public final class ConfigManager {
         CUSTOM_WIN_MULTIPLIER_ENABLED("custom-win-multiplier.enabled"),
         MONEY_ABBREVIATION_FORMAT_ENABLED("money-abbreviation-format.enabled");
 
+        private final RoulettePlugin plugin = JavaPlugin.getPlugin(RoulettePlugin.class);
         private final String path;
 
         Config(String path) {
@@ -176,11 +182,11 @@ public final class ConfigManager {
             return PluginUtils.translate(plugin.getConfig().getString(path));
         }
 
-        public List<String> asList() {
+        public @NotNull List<String> asList() {
             return PluginUtils.translate(plugin.getConfig().getStringList(path));
         }
 
-        public boolean asBoolean() {
+        public boolean asBool() {
             return plugin.getConfig().getBoolean(path);
         }
 
